requirements.txt
flask
python-docx
flask-cors
